requirements.txt
flask
requests
yt-dlp
playwright
torch
transformers
opencv-python
pillow
numpy
beautifulsoup4
python-docx
nltk
tqdm
mediapipe